+++ lib/image_processing/mini_magick.rb
@@ -113,7 +113,14 @@ module ImageProcessing
       end
 
       def apply_options(magick, define: {}, **options)
-        apply(magick, options)
+        options.each do |option, value|
+          case value
+          when true, nil then magick.send(option)
+          when false     then magick.send(option).+
+          else                magick.send(option, *value)
+          end
+        end
+
         define(magick, define)
       end